shift pytest support code into deltachat package so deltabot can make use of the test infrastructure
